+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Common:
 * Added dataset loader for Babel dataset.
 
+Task 3, Joint Embedding:
+* Metric: Added TotalVI metric 'latent_mixing_metric'.
+
 # neurips2021_multimodal_viash 0.4.0
 
 ## NEW FEATURES